Seperate out clip handle component
