+++ client/DAW/components/clip.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { connect } from 'react-redux';
 import { Button } from 'semantic-ui-react';
 import Draggable from 'react-draggable';
-import { Waveform } from '../components';
+import { ClipHandle, Waveform } from '../components';
 import { updateClipThunk, deleteClip } from '../project-store/reducers/clips';
 
 const styles = {
@@ -37,48 +37,7 @@ const styles = {
     margin: '1em',
     padding: '0.5em',
   },
-  clipHandle(side) {
-    return Object.assign({
-      position: 'absolute',
-      top: '0',
-      height: '100%',
-      width: '20px',
-      background: 'black',
-      cursor: 'col-resize',
-      opacity: '0.8',
-    }, side === 'left' ? { left: '0' } : { right: '0' });
-  },
 };
-
-class ClipHandle extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      x: 0,
-    };
-
-    this.handleDrag = this.handleDrag.bind(this);
-  }
-
-  handleDrag(e, data) {
-    this.props.handle(data.x);
-    this.setState({ x: 0 });
-  }
-
-  render() {
-    const { x } = this.state;
-    return (
-      <Draggable
-        axis="x"
-        onStart={e => e.stopPropagation()}
-        onDrag={this.handleDrag}
-        position={{ x, y: 0 }}
-      >
-        <div style={styles.clipHandle(this.props.side)} />
-      </Draggable>
-    );
-  }
-}
 
 class Clip extends React.Component {
   constructor(props) {
