+++ client/DAW/components/index.js
@@ -1,4 +1,5 @@
 export { default as Clip } from './clip';
+export { default as ClipHandle } from './clip-handle';
 export { default as DAW } from './DAW';
 export { default as FileItem } from './file-item';
 export { default as PlaybackControls } from './playback-controls';
